# specs/openapi/api.v1.yaml
openapi: "3.1.0"
info:
  title: "LongoX Platform API"
  version: "1.0.0"
  description: "LongoX API — low-code automation and internal tools platform"
  contact:
    name: "LongoX Support"
    url: "https://longox.io/support"
  license:
    name: "Proprietary"
servers:
  - url: "https://api.longox.io"
    description: "Production"
  - url: "https://staging-api.longox.io"
    description: "Staging"
  - url: "http://localhost:3000"
    description: "Development"
paths:
...
  /api/v1/ai/runs:
    post:
      operationId: createAiRun
      tags: ["AI"]
      summary: "Create an AI run"
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAiRunInput"
      responses:
        "201":
          description: "Created AI run"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiRun"
    get:
      operationId: listAiRuns
      tags: ["AI"]
      summary: "List AI runs"
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/PaginationPage"
        - $ref: "#/components/parameters/PaginationLimit"
      responses:
        "200":
          description: "List of AI runs"
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/AiRun"
                  total:
                    type: integer
...
  /api/v1/rag/knowledge-bases:
    post:
      operationId: createKnowledgeBase
      tags: ["RAG"]
      summary: "Create a knowledge base"
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateKnowledgeBaseInput"
      responses:
        "201":
          description: "Created knowledge base"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KnowledgeBase"
    get:
      operationId: listKnowledgeBases
      tags: ["RAG"]
      summary: "List knowledge bases"
      security:
        - BearerAuth: []
      responses:
        "200":
          description: "Knowledge bases"
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/KnowledgeBase"
  /api/v1/rag/knowledge-bases/{id}/documents:
    post:
      operationId: ingestDocument
      tags: ["RAG"]
      summary: "Ingest a document into a knowledge base"
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IngestDocumentInput"
      responses:
        "201":
          description: "Ingested document"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KnowledgeDocument"
  /api/v1/rag/knowledge-bases/{id}/search:
    post:
      operationId: searchKnowledgeBase
      tags: ["RAG"]
      summary: "Search a knowledge base"
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RagSearchInput"
      responses:
        "200":
          description: "Search results"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RagSearchResult"
  /api/v1/rag/knowledge-bases/{id}/query:
    post:
      operationId: queryKnowledgeBase
      tags: ["RAG"]
      summary: "Query a knowledge base with a question"
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RagQueryInput"
      responses:
        "200":
          description: "Query result"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RagQueryResult"

  # ─── Marketplace ────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "JWT Bearer token obtained from /auth/login or /auth/register"
  parameters:
...
    PaginationPage:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PaginationLimit:
      name: pageSize
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
  schemas:
...
    CreateAiRunInput:
      type: object
      required: [prompt]
      properties:
        modelId:
          type: string
        prompt:
          type: string
        systemPrompt:
          type: string
        temperature:
          type: number
          format: float
        maxTokens:
          type: integer
        guardrails:
          type: array
          items:
            type: string
        input:
          type: object
        metadata:
          type: object
    AiRun:
      type: object
      properties:
        id:
          type: string
        tenantId:
          type: string
        modelId:
          type: string
        status:
          type: string
        promptTokens:
          type: integer
        completionTokens:
          type: integer
        costUsd:
          type: number
          format: float
        result:
          type: string
          nullable: true
        errorMessage:
          type: string
          nullable: true
        guardrailHit:
          type: boolean
        startedAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
...
    CreateKnowledgeBaseInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        description:
          type: string
        embeddingModel:
          type: string
          default: text-embedding-3-small
        chunkingConfig:
          type: object
    KnowledgeBase:
      type: object
      properties:
        id:
          type: string
        tenantId:
          type: string
        name:
          type: string
        description:
          type: string
          nullable: true
        embeddingModel:
          type: string
        chunkingConfig:
          type: object
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    IngestDocumentInput:
      type: object
      required: [sourceRef]
      properties:
        sourceRef:
          type: string
        metadata:
          type: object
    KnowledgeDocument:
      type: object
      properties:
        id:
          type: string
        knowledgeBaseId:
          type: string
        sourceRef:
          type: string
        status:
          type: string
        chunkCount:
          type: integer
        metadata:
          type: object
        processedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    RagSearchInput:
      type: object
      required: [query]
      properties:
        query:
          type: string
        limit:
          type: integer
          default: 10
        minScore:
          type: number
          format: float
          default: 0.7
    RagSearchResult:
      type: object
      properties:
        results:
          type: array
          items:
            type: object
            properties:
              documentId:
                type: string
              content:
                type: string
              score:
                type: number
                format: float
              metadata:
                type: object
    RagQueryInput:
      type: object
      required: [question]
      properties:
        question:
          type: string
        model:
          type: string
        temperature:
          type: number
          format: float
        maxTokens:
          type: integer
    RagQueryResult:
      type: object
      properties:
        answer:
          type: string
        sources:
          type: array
          items:
            type: object
            properties:
              documentId:
                type: string
              content:
                type: string
              score:
                type: number
                format: float
...